+++ components/carousel/index.js
@@ -1,67 +1,60 @@
 import { Controller } from '@hotwired/stimulus'
+import { dataset } from '@newlogic-digital/utils-js'
 
 export class Carousel extends Controller {
-  static targets = ['content', 'counterMin', 'counterMax', 'pagination', 'progress', 'prev', 'next']
+  static targets = ['content', 'markerGroup', 'marker', 'prev', 'next']
 
-  paginationItemContent = '<div class="dot size-2 bg-body-secondary transition data-active:bg-accent cursor-pointer"></div>'
-
-  async connect() {
-    await this.scroll()
-
-    if (this.hasPaginationTarget) {
-      const { paginationCarousel } = await import('winduum/src/components/carousel/index.js')
-
-      paginationCarousel(this.contentTarget, {
-        element: this.paginationTarget,
-        itemContent: this.paginationItemContent,
-      })
-    }
+  static values = {
+    vertical: Boolean,
   }
 
-  async scroll() {
-    const { scrollCarousel } = await import('winduum/src/components/carousel/index.js')
+  async connect() {
+    const { setSnappedAttribute, toggleScrollState } = await import('winduum/src/components/carousel/index.js')
 
-    this.toggleScrollAttributes()
+    this.abortController = new AbortController()
+    const signal = this.abortController.signal
 
-    scrollCarousel(this.contentTarget, {
-      counterMinElement: this.hasCounterMinTarget && this.counterMinTarget,
-      counterMaxElement: this.hasCounterMaxTarget && this.counterMaxTarget,
-      progressElement: this.hasProgressTarget && this.progressTarget,
-      pagination: {
-        element: this.hasPaginationTarget && this.paginationTarget,
-      },
+    this.contentTarget.addEventListener('scrollsnapchanging', (event) => {
+      setSnappedAttribute(this.contentTarget, event.snapTargetInline ?? event.snapTargetBlock, this.hasMarkerGroupTarget ? this.markerGroupTarget : null)
+    }, { signal })
+
+    this.contentTarget.addEventListener('scroll', () => {
+      toggleScrollState(this.contentTarget, {
+        prevElement: this.hasPrevTarget ? this.prevTarget : null,
+        nextElement: this.hasNextTarget ? this.nextTarget : null,
+        vertical: this.verticalValue,
+      })
+    }, { signal })
+  }
+
+  disconnect() {
+    this.abortController?.abort()
+  }
+
+  markerTargetConnected(element) {
+    dataset(element, 'action').add(`click->${this.identifier}#scrollToMarker:prevent`)
+  }
+
+  async scrollToMarker({ currentTarget }) {
+    const { scrollToMarker } = await import('winduum/src/components/carousel/index.js')
+
+    scrollToMarker(this.contentTarget, currentTarget, this.markerGroupTarget, this.verticalValue ? { block: 'start' } : {})
+  }
+
+  async scroll(direction) {
+    const { scrollBy } = await import('winduum/src/components/carousel/index.js')
+
+    scrollBy(this.contentTarget, {
+      direction,
+      vertical: this.verticalValue,
     })
   }
 
-  toggleScrollAttributes() {
-    const scrollStart = this.contentTarget.scrollLeft <= 0
-    const scrollEnd = this.contentTarget.scrollLeft >= this.contentTarget.scrollWidth - this.contentTarget.clientWidth
-    const scrollNone = this.contentTarget.scrollWidth - this.contentTarget.clientWidth === 0
-
-    if (this.hasPrevTarget && this.hasNextTarget) {
-      this.prevTarget.disabled = scrollStart
-      this.nextTarget.disabled = scrollEnd
-    }
-
-    this.element.toggleAttribute('data-scroll-start', scrollStart)
-    this.element.toggleAttribute('data-scroll-end', scrollEnd)
-    this.element.toggleAttribute('data-scroll-none', scrollNone)
-  }
-
   scrollPrev() {
-    this.contentTarget.scroll({ left: this.contentTarget.scrollLeft - this.contentTarget.children[0].offsetWidth })
+    this.scroll(-1)
   }
 
   scrollNext() {
-    this.contentTarget.scroll({ left: this.contentTarget.scrollLeft + this.contentTarget.children[0].offsetWidth })
-  }
-
-  async scrollTo({ currentTarget }) {
-    const { scrollTo } = await import('winduum/src/components/carousel/index.js')
-
-    const siblingElements = [...currentTarget.parentElement.children]
-    const index = siblingElements.indexOf(currentTarget)
-
-    scrollTo(this.contentTarget, index)
+    this.scroll(1)
   }
 }