+++ components/control/index.js
@@ -1,7 +1,22 @@
 import { Controller } from '@hotwired/stimulus'
-import { dispatchCustomEvent } from '@newlogic-digital/utils-js'
+import { dataset, dispatchCustomEvent } from '@newlogic-digital/utils-js'
 
 export class Control extends Controller {
+  activeAttribute = 'data-active'
+
+  connect() {
+    this.toggleActiveAttribute()
+    dataset(this.element, 'action').add(`change->${this.identifier}#toggleActiveAttribute`)
+  }
+
+  toggleActiveAttribute() {
+    const telCountryCode = this.element.querySelector('[autocomplete="tel-country-code"]')
+
+    if (telCountryCode) telCountryCode.dataset.value = telCountryCode.value
+
+    this.element.toggleAttribute(this.activeAttribute, !!this.element.querySelector('input:not([type="hidden"]), textarea, select')?.value)
+  }
+
   stepUp() {
     this.element.querySelector('input').stepUp()
     dispatchCustomEvent(this.element.querySelector('input'))